+++ extendr-api/src/prelude.rs
@@ -34,7 +34,7 @@ pub use crate::{
 
 pub use super::wrapper::{
     AltComplexImpl, AltIntegerImpl, AltLogicalImpl, AltRawImpl, AltRealImpl, AltStringImpl, Altrep,
-    AltrepImpl, RArray, RColumn, RMatrix, RMatrix3D, RMatrix4D, RMatrix5D,
+    AltrepImpl, ColIter, RArray, RColumn, RMatrix, RMatrix3D, RMatrix4D, RMatrix5D, RowIter,
 };
 
 #[cfg(use_r_altlist)]